% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate with Semestry Termtime API}
\usage{
authenticate(api_key, base_url)
}
\arguments{
\item{api_key}{The API key for Semestry.}

\item{base_url}{The base URL of the Semestry API.}
}
\value{
An authenticated Semestry object.
}
\description{
This function authenticates with the Semestry Termtime API by verifying the API key and base URL.
It returns an authenticated Semestry object that can be used for subsequent API calls.
}
