% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activities.R
\name{get_activities}
\alias{get_activities}
\title{Retrieve activity data from the Semestry API}
\usage{
get_activities(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved activity data from the API.
}
\description{
This function sends a GET request to the activities endpoint of the Semestry API and retrieves the activity data.
}
