% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activities_report_attr.R
\name{get_activities_report_attr}
\alias{get_activities_report_attr}
\title{Retrieve activity report attr from the Semestry API}
\usage{
get_activities_report_attr(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved activity report data from the API.
}
\description{
This function sends a GET request to the report room endpoint of the Semestry API and retrieves the room booking data.
}
