% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity_groups.R
\name{get_activity_groups}
\alias{get_activity_groups}
\title{Retrieve activity group data from the Semestry API}
\usage{
get_activity_groups(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved activity group data from the API.
}
\description{
This function sends a GET request to the activity groups endpoint of the Semestry API and retrieves the activity group data.
}
