% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assessment_types.R
\name{get_assessment_types}
\alias{get_assessment_types}
\title{Retrieve assessment type data from the Semestry API}
\usage{
get_assessment_types(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved assessment type data from the API.
}
\description{
This function sends a GET request to the assessment types endpoint of the Semestry API and retrieves the assessment type data.
}
