% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_campuses.R
\name{get_campuses}
\alias{get_campuses}
\title{Retrieve campus data from the Semestry API}
\usage{
get_campuses(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved campus data from the API.
}
\description{
This function sends a GET request to the campuses endpoint of the Semestry API and retrieves the campus data.
}
