% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capabilities.R
\name{get_capabilities}
\alias{get_capabilities}
\title{Retrieve capability data from the Semestry API}
\usage{
get_capabilities(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved capability data from the API.
}
\description{
This function sends a GET request to the capabilities endpoint of the Semestry API and retrieves the capability data.
}
