% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contact_requirements.R
\name{get_contact_requirements}
\alias{get_contact_requirements}
\title{Retrieve contact requirement data from the Semestry API}
\usage{
get_contact_requirements(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved contact requirement data from the API.
}
\description{
This function sends a GET request to the contact requirements endpoint of the Semestry API and retrieves the contact requirement data.
}
