% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_courses.R
\name{get_courses}
\alias{get_courses}
\title{Retrieve course data from the Semestry API}
\usage{
get_courses(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved course data from the API.
}
\description{
This function sends a GET request to the courses endpoint of the Semestry API and retrieves the course data.
}
