% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_departments.R
\name{get_departments}
\alias{get_departments}
\title{Retrieve department data from the Semestry API}
\usage{
get_departments(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved department data from the API.
}
\description{
This function sends a GET request to the departments endpoint of the Semestry API and retrieves the department data.
}
