% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_floors.R
\name{get_floors}
\alias{get_floors}
\title{Retrieve floor data from the Semestry API}
\usage{
get_floors(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved floor data from the API.
}
\description{
This function sends a GET request to the floors endpoint of the Semestry API and retrieves the floor data.
}
