% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modules.R
\name{get_modules}
\alias{get_modules}
\title{Retrieve module data from the Semestry API}
\usage{
get_modules(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved module data from the API.
}
\description{
This function sends a GET request to the modules endpoint of the Semestry API and retrieves the module data.
}
