% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offers.R
\name{get_offers}
\alias{get_offers}
\title{Retrieve offers from the Semestry API}
\usage{
get_offers(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved offers from the API.
}
\description{
This function sends a GET request to the offers endpoint of the Semestry API and retrieves the offers.
}
\examples{
\dontrun{
  semestry <- authenticate_semestry()
  get_offers(semestry)
}

}
