% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ping.R
\name{get_ping}
\alias{get_ping}
\title{Ping the Semestry API}
\usage{
get_ping(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The ping response from the API.
}
\description{
This function sends a GET request to the ping endpoint of the Semestry API to check connectivity.
}
