% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_booking.R
\name{get_report_booking}
\alias{get_report_booking}
\title{Retrieve booking report from the Semestry API}
\usage{
get_report_booking(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved report booking data from the API.
}
\description{
This function sends a GET request to the report booking endpoint of the Semestry API and retrieves the booking booking data.
}
