% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_room_relationships.R
\name{get_room_relationships}
\alias{get_room_relationships}
\title{Retrieve room relationship data from the Semestry API}
\usage{
get_room_relationships(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved room relationship data from the API.
}
\description{
This function sends a GET request to the room relationships endpoint of the Semestry API and retrieves the room relationship data.
}
