% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roombookings.R
\name{get_roombookings}
\alias{get_roombookings}
\title{Retrieve room booking data from the Semestry API}
\usage{
get_roombookings(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved room booking data from the API.
}
\description{
This function sends a GET request to the roombookings endpoint of the Semestry API and retrieves the room booking data.
}
