% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schedule_rooms.R
\name{get_schedule_rooms}
\alias{get_schedule_rooms}
\title{Retrieve schedule rooms data from the Semestry API}
\usage{
get_schedule_rooms(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved schedule rooms data from the API.
}
\description{
This function sends a GET request to the schedule/rooms endpoint of the Semestry API and retrieves the schedule rooms data.
}
