% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_staff.R
\name{get_staff}
\alias{get_staff}
\title{Retrieve staff data from the Semestry API}
\usage{
get_staff(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved staff data from the API.
}
\description{
This function sends a GET request to the staff endpoint of the Semestry API and retrieves the staff data.
}
\examples{
\dontrun{
  semestry <- authenticate_semestry()
  get_staff(semestry)
}

}
