% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_staff_groups.R
\name{get_staff_groups}
\alias{get_staff_groups}
\title{Retrieve staff group data from the Semestry API}
\usage{
get_staff_groups(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved staff group data from the API.
}
\description{
This function sends a GET request to the staff groups endpoint of the Semestry API and retrieves the staff group data.
}
