% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_staff_pools.R
\name{get_staff_pools}
\alias{get_staff_pools}
\title{Retrieve staff pools from the Semestry API}
\usage{
get_staff_pools(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved staff pools from the API.
}
\description{
This function sends a GET request to the staffpools endpoint of the Semestry API and retrieves the staff pools.
}
