% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_group_maps.R
\name{get_student_group_maps}
\alias{get_student_group_maps}
\title{Retrieve student group mapppings from the Semestry API}
\usage{
get_student_group_maps(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved student group maps from the API.
}
\description{
This function sends a GET request to the studentgroupmaps endpoint of the Semestry API and retrieves the student group mappings data.
}
