% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_groups.R
\name{get_student_groups}
\alias{get_student_groups}
\title{Retrieve student group data from the Semestry API}
\usage{
get_student_groups(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved student group data from the API.
}
\description{
This function sends a GET request to the student groups endpoint of the Semestry API and retrieves the student group data.
}
