% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students.R
\name{get_students}
\alias{get_students}
\title{Retrieve student data from the Semestry API}
\usage{
get_students(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved student data from the API.
}
\description{
This function sends a GET request to the students endpoint of the Semestry API and retrieves the student data.
}
\examples{
\dontrun{
  semestry <- authenticate_semestry()
  get_students(semestry)
}

}
