% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supergroups.R
\name{get_supergroups}
\alias{get_supergroups}
\title{Retrieve supergroup data from the Semestry API}
\usage{
get_supergroups(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved supergroup data from the API.
}
\description{
This function sends a GET request to the supergroups endpoint of the Semestry API and retrieves the supergroup data.
}
