% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeframes.R
\name{get_timeframes}
\alias{get_timeframes}
\title{Retrieve timeframes data from the Semestry API}
\usage{
get_timeframes(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved timeframes data from the API.
}
\description{
This function sends a GET request to the timeframes endpoint of the Semestry API and retrieves the timeframes data.
}
