% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_webhooks.R
\name{get_webhooks}
\alias{get_webhooks}
\title{Retrieve webhook data from the Semestry API}
\usage{
get_webhooks(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved webhook data from the API.
}
\description{
This function sends a GET request to the webhooks endpoint of the Semestry API and retrieves the webhook data.
}
