% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rooms.R
\name{get_rooms}
\alias{get_rooms}
\title{Retrieve room data from the Semestry API}
\usage{
get_rooms(semestry, timeout = 30)
}
\arguments{
\item{semestry}{An authenticated Semestry object.}

\item{timeout}{The timeout duration for the GET request (default: 30 seconds).}
}
\value{
The retrieved room data from the API.
}
\description{
This function sends a GET request to the rooms endpoint of the Semestry API and retrieves the room data.
}
\examples{
\dontrun{
  semestry <- authenticate_semestry()
  get_rooms(semestry)
}

}
